+++ Libplanet.Tests/Action/ActionContextTest.cs
@@ -150,7 +150,7 @@ namespace Libplanet.Tests.Action
         {
             IKeyValueStore keyValueStore = new MemoryKeyValueStore();
             ITrie previousBlockStatesTrie = new MerkleTrie(keyValueStore);
-            previousBlockStatesTrie = previousBlockStatesTrie.Set(new byte[0], Null.Value);
+            previousBlockStatesTrie = previousBlockStatesTrie.Set(default, Null.Value);
             var actionContext = new ActionContext(
                 signer: _address,
                 txid: _txid,